change main file structure
